rename file so it isnt the same as a module name (mujoco)
